fix: scan-build findings — MallocOverflow cast and dead store
IccHeuristicsDataValidation.cpp:447 — cast scanLimit to size_t before
multiplication to avoid alpha.security.MallocOverflow (scanLimit is capped
to 4096 so no actual overflow, but the cast makes intent explicit).

IccHeuristicsRawPost.cpp:1106 — hasSvcn was assigned but never read.
Added spectral tag consistency check: svcn without sdin is suspicious.

+++ iccanalyzer-lite/IccHeuristicsDataValidation.cpp
@@ -444,7 +444,8 @@ printf("[H66] Comprehensive NumArray NaN/Inf Scan\n");
     if (nVals == 0 || nVals > 1048576) continue; // skip empty or huge
 
     icUInt32Number scanLimit = (nVals > 4096) ? 4096 : nVals;
-    icFloatNumber *vals = (icFloatNumber*)malloc(scanLimit * sizeof(icFloatNumber));
+    // Guard against overflow: scanLimit <= 4096 so product fits in uint32
+    icFloatNumber *vals = (icFloatNumber*)malloc(static_cast<size_t>(scanLimit) * sizeof(icFloatNumber));
     if (!vals) continue;
 
     if (numArr->GetValues(vals, 0, scanLimit)) {
